TEMP: We're nudging towards a solution
Note for the future:
The Setup project has dependencies to Services
and Connections. Storage.MongoDB has dependency to
Setup.

These should be somehow broken.

One possible solution:
The Orleans.InProcess could include Services and Setup .cs files for compilation. The Services C# types would need to be
internalized. Either we do it explicitly in code, or we create a Roslyn source generator that can be configured to include
files from a folder and then generate files where public is switched to internal.